Move the matchers to `active_record/acts_as/matchers`
This is for consistency with other gems, like rspec-rails, and shoulda-matchers.
