+++ spec/rspec_matchers_spec.rb
@@ -1,5 +1,5 @@
 require 'models'
-require 'rspec/acts_as_matchers'
+require 'active_record/acts_as/matchers'
 
 RSpec.describe "Custom RSpec matchers" do
   it { expect(Product).to be_actable }
